Remove jsbeautifier from css install time deps
